Convert Days 2017 to build with Bazel
This is the initial change to get each of the days to build with Bazel.  This does not split out tests into their own libraries, but that would be useful work in the future.

Many of the C++ files got autoformatted as part of this change - apologies for the confusing diff.